+++ src/ios/Diagnostic.h
@@ -26,6 +26,9 @@
 #import <arpa/inet.h> // For AF_INET, etc.
 #import <ifaddrs.h> // For getifaddrs()
 #import <net/if.h> // For IFF_LOOPBACK
+#import <mach/machine.h>
+#import <sys/types.h>
+#import <sys/sysctl.h>
 
 @interface Diagnostic : CDVPlugin <CBCentralManagerDelegate, CLLocationManagerDelegate>
 
@@ -87,5 +90,7 @@
 - (void) isMotionRequestOutcomeAvailable: (CDVInvokedUrlCommand*)command;
 - (void) getMotionAuthorizationStatus: (CDVInvokedUrlCommand*)command;
 - (void) requestMotionAuthorization: (CDVInvokedUrlCommand*)command;
+
+- (void) getArchitecture: (CDVInvokedUrlCommand*)command;
 
 @end