+++ www/ios/diagnostic.js
@@ -76,6 +76,15 @@ var Diagnostic = (function(){
         "NOT_DETERMINED": "not_determined" // Motion authorization request status outcome cannot be determined on device
     };
 
+    Diagnostic.cpuArchitecture = {
+        UNKNOWN: "unknown",
+        ARMv6: "ARMv6",
+        ARMv7: "ARMv7",
+        ARMv8: "ARMv8",
+        X86: "X86",
+        X86_64: "X86_64"
+    };
+
     // Placeholder listeners
     Diagnostic._onBluetoothStateChange =
         Diagnostic._onLocationStateChange = function(){};
@@ -955,6 +964,22 @@ var Diagnostic = (function(){
             errorCallback,
             'Diagnostic',
             'getMotionAuthorizationStatus',
+            []);
+    };
+
+    /**
+     * Returns CPU architecture of the current device.
+     *
+     * @param {Function} successCallback -  The callback which will be called when the operation is successful.
+     * This callback function is passed a single string parameter defined as a constant in `cordova.plugins.diagnostic.cpuArchitecture`.
+     * @param {Function} errorCallback -  The callback which will be called when the operation encounters an error.
+     *  This callback function is passed a single string parameter containing the error message.
+     */
+    Diagnostic.getArchitecture = function(successCallback, errorCallback) {
+        return cordova.exec(successCallback,
+            errorCallback,
+            'Diagnostic',
+            'getArchitecture',
             []);
     };